Se termina la primera fase del proyecto

+++ coreComercios/models.py
@@ -1,12 +1,11 @@
 import os
+from django_resized import ResizedImageField
 from datetime import datetime
 from django.db import models
 from django.contrib.auth.models import User
 
 #funcion de carga de imagen de comercio
 def custom_upload_ComercioSup(instance, filename):
-    old_instance = Comercio.objects.get(pk=instance.pk)
-
     # Genera el momento de carga de imagen
     timestamp = datetime.timestamp(datetime.now())
 
@@ -17,12 +16,14 @@ def custom_upload_ComercioSup(instance, filename):
     #ruta donde se va a guardar
     ruta = "comercios/{}.{}".format(timestamp, datosImagen[0])
 
-    old_instance.img_superior.delete()
+    # Borra la imagen anterior en caso de que exista
+    if instance.pk is not None:
+        old_instance = Comercio.objects.get(pk=instance.pk)
+        old_instance.img_superior.delete()
+
     return ruta
 
 def custom_upload_Comercio(instance, filename):
-    old_instance = Comercio.objects.get(pk=instance.pk)
-
     # Genera el momento de carga de imagen
     timestamp = datetime.timestamp(datetime.now())
 
@@ -33,13 +34,15 @@ def custom_upload_Comercio(instance, filename):
     #ruta donde se va a guardar
     ruta = "comercios/{}.{}".format(timestamp, datosImagen[0])
 
-    old_instance.img_acercade.delete()
+    # Borra la imagen anterior en caso de que exista
+    if instance.pk is not None:
+        old_instance = Comercio.objects.get(pk=instance.pk)
+        old_instance.img_acercade.delete()
+
     return ruta
 
 #funcion de carga de imagen de productos
 def custom_upload_Producto(instance, filename):
-    old_instance = ImagenesProducto.objects.get(pk=instance.pk)
-
     # Genera el momento de carga de imagen
     timestamp = datetime.timestamp(datetime.now())
 
@@ -50,7 +53,11 @@ def custom_upload_Producto(instance, filename):
     #ruta donde se va a guardar
     ruta = "productos/{}.{}".format(timestamp, datosImagen[0])
 
-    old_instance.imagen.delete()
+    # Borra la imagen anterior en caso de que exista
+    if instance.pk is not None:
+        old_instance = ImagenesProducto.objects.get(pk=instance.pk)
+        old_instance.imagen.delete()
+
     return ruta
 
 # Create your models here.
@@ -62,8 +69,8 @@ class Comercio(models.Model):
     descripcion     = models.TextField(verbose_name="Descripción", default="")
     redessociales   = models.TextField(verbose_name="Redes Sociales")
     contacto        = models.TextField(verbose_name="Contácto")
-    img_superior    = models.ImageField(upload_to=custom_upload_ComercioSup, null=False, default="comercios/noimageSup.jpg", verbose_name="Imagen superior")
-    img_acercade    = models.ImageField(upload_to=custom_upload_Comercio, null=False, default="comercios/noimageAbout.jpg", verbose_name="Imagen de acerca de")
+    img_superior    = ResizedImageField(upload_to=custom_upload_ComercioSup, size=[900, 600], null=False, default="comercios/noimageSup.jpg", verbose_name="Imagen superior")
+    img_acercade    = ResizedImageField(upload_to=custom_upload_Comercio, size=[500, 433], null=False, default="comercios/noimageAbout.jpg", verbose_name="Imagen de acerca de")
     
     def __str__(self):
         return self.nombre
@@ -81,9 +88,9 @@ class Producto(models.Model):
 
 class ImagenesProducto(models.Model):
     producto    = models.ForeignKey(Producto, verbose_name="Producto asociado", on_delete=models.CASCADE)
-    imagen      = models.ImageField(upload_to=custom_upload_Producto, null=False, verbose_name="Imagen del producto", default="productos/noimageProd.jpg")
+    imagen      = ResizedImageField(upload_to=custom_upload_Producto, size=[600, 600], null=False, default="productos/noimageProd.jpg", verbose_name="Imagen del producto")
     principal   = models.BooleanField(verbose_name="Imagen principal", default=False)
     estado      = models.BooleanField(verbose_name="Estado", default=False)
     
     def __str__(self):
-        return self.imagen
+        return self.imagen.url
